First attempt at day3p2

+++ day3/src/main.cpp
@@ -17,10 +17,12 @@ int main(int argc, const char* argv[]) {
     std::vector<std::string> fileOutput = fh.readFile(argv[1]);
     
     int result = bb.solve(fileOutput);
-    int result2 = bb.solve2(fileOutput);
+    long result2 = bb.solve2(fileOutput);
 
     std::cout << "Result 1: " << std::to_string(result) << std::endl;
     std::cout << "Result 2: " << std::to_string(result2) << std::endl;
+
+    // 776380 is too low
 
     return 0;
 }
